+++ ouroboros-consensus-cardano/ouroboros-consensus-cardano.cabal
@@ -33,6 +33,11 @@ library
                        Ouroboros.Consensus.Cardano.CanHardFork
                        Ouroboros.Consensus.Cardano.Node
                        Ouroboros.Consensus.Cardano.ShelleyBased
+                       Ouroboros.Consensus.Example
+                       Ouroboros.Consensus.Example.ShelleyBased
+                       Ouroboros.Consensus.Example.Block
+                       Ouroboros.Consensus.Example.Eras
+                       Ouroboros.Consensus.Example.Node
 
   build-depends:       base              >=4.9   && <4.15
                      , bytestring        >=0.10  && <0.11
@@ -58,7 +63,6 @@ library
                      , ouroboros-consensus
                      , ouroboros-consensus-byron
                      , ouroboros-consensus-shelley
-                     , ouroboros-consensus-example
 
   default-language:    Haskell2010
   ghc-options:         -Wall
