+++ ouroboros-consensus-cardano/src/Ouroboros/Consensus/Cardano.hs
@@ -55,6 +55,9 @@ import           Ouroboros.Consensus.Shelley.Ledger
 import           Ouroboros.Consensus.Shelley.Node as X
 
 import qualified Ouroboros.Consensus.Example as Example
+import qualified Ouroboros.Consensus.Example.Eras as Example
+import qualified Ouroboros.Consensus.Example.Node as Example
+import           Ouroboros.Consensus.Example.Node as X
 
 import           Ouroboros.Consensus.Cardano.Block
 import           Ouroboros.Consensus.Cardano.ByronHFC
@@ -126,13 +129,14 @@ data Protocol (m :: Type -> Type) blk p where
     -> Example.ProtocolParamsExample
     -> ProtocolParamsTransition
          (ShelleyBlock StandardShelley)
-         (ShelleyBlock StandardExample)
+         (ShelleyBlock Example.StandardExample)
     -> Protocol m (Example.ExampleBlock StandardCrypto) Example.ProtocolExample
 
 verifyProtocol :: Protocol m blk p -> (p :~: BlockProtocol blk)
 verifyProtocol ProtocolByron{}   = Refl
 verifyProtocol ProtocolShelley{} = Refl
 verifyProtocol ProtocolCardano{} = Refl
+verifyProtocol ProtocolExample{} = Refl
 
 {-------------------------------------------------------------------------------
   Data required to run a protocol
@@ -166,6 +170,17 @@ protocolInfo (ProtocolCardano
       paramsShelleyAllegra
       paramsAllegraMary
 
+protocolInfo (ProtocolExample
+               paramsShelleyBased
+               paramsShelley
+               paramsExample
+               paramsShelleyExample) =
+    protocolInfoExample
+      paramsShelleyBased
+      paramsShelley
+      paramsExample
+      paramsShelleyExample
+
 {-------------------------------------------------------------------------------
   Evidence that we can run all the supported protocols
 -------------------------------------------------------------------------------}
@@ -174,6 +189,7 @@ runProtocol :: Protocol m blk p -> Dict (RunNode blk)
 runProtocol ProtocolByron{}   = Dict
 runProtocol ProtocolShelley{} = Dict
 runProtocol ProtocolCardano{} = Dict
+runProtocol ProtocolExample{} = Dict
 
 {-------------------------------------------------------------------------------
   Client support for the protocols: what you need as a client of the node
@@ -202,17 +218,25 @@ data ProtocolClient blk p where
          (CardanoBlock StandardCrypto)
          ProtocolCardano
 
+  ProtocolClientExample
+    :: EpochSlots
+    -> ProtocolClient
+         (Example.ExampleBlock StandardCrypto)
+         Example.ProtocolExample
+
 -- | Sanity check that we have the right type combinations
 verifyProtocolClient :: ProtocolClient blk p -> (p :~: BlockProtocol blk)
 verifyProtocolClient ProtocolClientByron{}   = Refl
 verifyProtocolClient ProtocolClientShelley{} = Refl
 verifyProtocolClient ProtocolClientCardano{} = Refl
+verifyProtocolClient ProtocolClientExample{} = Refl
 
 -- | Sanity check that we have the right class instances available
 runProtocolClient :: ProtocolClient blk p -> Dict (RunNode blk)
 runProtocolClient ProtocolClientByron{}   = Dict
 runProtocolClient ProtocolClientShelley{} = Dict
 runProtocolClient ProtocolClientCardano{} = Dict
+runProtocolClient ProtocolClientExample{} = Dict
 
 -- | Data required by clients of a node running the specified protocol.
 protocolClientInfo :: ProtocolClient blk p -> ProtocolClientInfo blk
@@ -224,3 +248,6 @@ protocolClientInfo ProtocolClientShelley =
 
 protocolClientInfo (ProtocolClientCardano epochSlots) =
     protocolClientInfoCardano epochSlots
+
+protocolClientInfo (ProtocolClientExample epochSlots) =
+    protocolClientInfoExample epochSlots
