+++ ouroboros-consensus-cardano/src/Ouroboros/Consensus/Example.hs
@@ -1,0 +1,46 @@
+{-# LANGUAGE DataKinds           #-}
+{-# LANGUAGE GADTs               #-}
+{-# LANGUAGE KindSignatures      #-}
+{-# LANGUAGE ScopedTypeVariables #-}
+{-# LANGUAGE TypeOperators       #-}
+
+module Ouroboros.Consensus.Example (
+    -- * The block type of the Cardano block chain
+    ExampleBlock
+    -- * Supported protocols
+  , ProtocolExample
+  ) where
+
+import           Data.Kind (Type)
+import           Data.Type.Equality
+
+import           Cardano.Chain.Slotting (EpochSlots)
+
+import           Ouroboros.Consensus.Block
+import           Ouroboros.Consensus.Node.ProtocolInfo
+import           Ouroboros.Consensus.Node.Run
+import           Ouroboros.Consensus.Protocol.Abstract as X
+import           Ouroboros.Consensus.Protocol.PBFT as X
+import           Ouroboros.Consensus.Util
+import           Ouroboros.Consensus.Util.IOLike
+
+import           Ouroboros.Consensus.HardFork.Combinator
+
+import           Ouroboros.Consensus.Shelley.Ledger
+import           Ouroboros.Consensus.Shelley.Node as X
+
+import           Ouroboros.Consensus.Example.Block
+import           Ouroboros.Consensus.Example.Node
+
+{-------------------------------------------------------------------------------
+  Supported protocols
+
+  We list these as explicit definitions here (rather than derived through
+  'BlockProtocol'), and then /verify/ in 'verifyProtocol' that these definitions
+  match. This provides an additional sanity check that we are not accidentally
+  breaking any assumptions made in @cardano-node@.
+-------------------------------------------------------------------------------}
+
+type ProtocolExample = HardForkProtocol '[ ShelleyBlock StandardShelley
+                                         , ShelleyBlock StandardExample
+                                         ]
