+++ ouroboros-consensus-cardano/src/Ouroboros/Consensus/Example/Block.hs
