+++ ouroboros-consensus-cardano/src/Ouroboros/Consensus/Example/Node.hs
@@ -41,6 +41,7 @@ import           Cardano.Prelude (cborError)
 
 import           Ouroboros.Consensus.Block
 import           Ouroboros.Consensus.Cardano.CanHardFork
+import           Ouroboros.Consensus.Cardano.Node (ProtocolParamsTransition (..))
 import           Ouroboros.Consensus.Config
 import qualified Ouroboros.Consensus.HardFork.History as History
 import           Ouroboros.Consensus.HeaderValidation
@@ -77,8 +78,6 @@ import           Ouroboros.Consensus.Example.ShelleyBased
 {-------------------------------------------------------------------------------
   SerialiseHFC
 -------------------------------------------------------------------------------}
-
-instance ShelleyBasedEra era => SerialiseConstraintsHFC (ShelleyBlock era)
 
 instance ExampleHardForkConstraints c => SerialiseHFC (ExampleEras c) where
   encodeDiskHfcBlock (ExampleCodecConfig ccfgShelley ccfgExample) = \case
@@ -189,14 +188,6 @@ instance ExampleHardForkConstraints c
 {-------------------------------------------------------------------------------
   ProtocolInfo
 -------------------------------------------------------------------------------}
-
--- | Parameters needed to transition between two eras.
---
--- The two eras are phantom type parameters of this type to avoid mixing up
--- multiple 'ProtocolParamsTransition's
-data ProtocolParamsTransition eraFrom eraTo = ProtocolParamsTransition {
-      transitionTrigger    :: TriggerHardFork
-    }
 
 -- | Parameters needed to run Shelley
 data ProtocolParamsExample = ProtocolParamsExample {
