--- ouroboros-consensus-example/README.md
@@ -1,8 +1,0 @@
-# ouroboros-consensus-cardano
-
-This package contains:
-
-* `src` : Support for Hard Fork between Cardano eras.
-
-* `tools/db-analyser`: tool to analyse or validate a database containing blocks
-  of some Cardano era.
