--- ouroboros-consensus-example/ouroboros-consensus-example.cabal
@@ -1,68 +1,0 @@
-name:                  ouroboros-consensus-example
-version:               0.1.0.0
-synopsis:              An instantation of the Ouroboros consensus layer intended as a template for prototype consensus protocols based on Shelley
--- description:
-license:               Apache-2.0
-license-files:
-  LICENSE
-  NOTICE
-copyright:             2020 Input Output (Hong Kong) Ltd.
-author:                IOHK Engineering Team
-maintainer:            operations@iohk.io
-category:              Network
-build-type:            Simple
-cabal-version:         >=1.10
-
-source-repository head
-  type:     git
-  location: https://github.com/input-output-hk/ouroboros-network
-
-flag asserts
-  description: Enable assertions
-  manual:      False
-  default:     False
-
-library
-  hs-source-dirs:      src
-  exposed-modules:
-                       Ouroboros.Consensus.Example
-                       Ouroboros.Consensus.Example.Block
-                       Ouroboros.Consensus.Example.Eras
-                       Ouroboros.Consensus.Example.ShelleyHFC
-                       Ouroboros.Consensus.Example.Node
-                       Ouroboros.Consensus.Example.ShelleyBased
-
-  build-depends:       base              >=4.9   && <4.15
-                     , bytestring        >=0.10  && <0.11
-                     , cborg             >=0.2.2 && <0.3
-                     , containers        >=0.5   && <0.7
-                     , mtl               >=2.2   && <2.3
-                     , nothunks
-                     , serialise         >=0.2   && <0.3
-                     , text              >=1.2   && <1.3
-                     , these             >=1.1   && <1.2
-
-                     , cardano-binary
-                     , cardano-crypto-class
-                     , cardano-crypto-wrapper
-                     , cardano-ledger
-                     , cardano-prelude
-                     , cardano-slotting
-                     , shelley-spec-ledger
-                     , cardano-ledger-shelley-ma
-                     , cardano-ledger-example-shelley
-                     , ouroboros-network
-                     , ouroboros-consensus
-                     , ouroboros-consensus-shelley
-
-  default-language:    Haskell2010
-  ghc-options:         -Wall
-                       -Wcompat
-                       -Wincomplete-uni-patterns
-                       -Wincomplete-record-updates
-                       -Wpartial-fields
-                       -Widentities
-                       -Wredundant-constraints
-                       -Wmissing-export-lists
-  if flag(asserts)
-    ghc-options:       -fno-ignore-asserts
